requirements.txt
fastapi
uvicorn
llama-index
chromadb
google-cloud-storage
openai
sentence-transformers
python-dotenv
--extra-index-url https://download.pytorch.org/whl/cpu
torch==2.1.0+cpu
transformers